feat: create alias on migrate if one doesn't exist
